Moved type2name to utils
