+++ src/util/CMakeLists.txt
@@ -32,7 +32,7 @@ add_library(util_esbmc xml_irep.cpp xml.cpp
         type_byte_size.cpp goto_expr_factory.cpp
         string_constant.cpp c_types.cpp ieee_float.cpp c_qualifiers.cpp
         c_sizeof.cpp c_link.cpp c_typecast.cpp fix_symbol.cpp destructor.cpp
-        c_expr2string.cpp cpp_expr2string.cpp
+        c_expr2string.cpp cpp_expr2string.cpp type2name.cpp
         message.cpp
         )
 # Boost is needed by anything that touches irep2
